chatbot/requirements.txt
fastapi
uvicorn
openai
google-generativeai
supabase
python-dotenv
pydantic
pydantic-settings
PyJWT
